fix(web/applications): rename middleware (boas-1661)
