+++ apps/web/src/server/applications/case/representations/representation/address-details/address-details.router.js
@@ -7,7 +7,7 @@ import {
 import { addressDetailsValidation } from './address-details.validators.js';
 import { addQueryToLocals, addRepresentationToLocals } from '../representation.middleware.js';
 import { repRoutes } from '../utils/get-representation-page-urls.js';
-import { registerCaseId } from '../../applications-relevant-reps.locals.js';
+import { registerCaseId } from '../../../../create-new-case/applications-create.locals.js';
 
 const relevantRepAddressDetailsRouter = createRouter({ mergeParams: true });
