+++ apps/web/src/server/applications/create-new-case/applications-create.locals.js
@@ -7,15 +7,3 @@ export const registerCaseId = ({ params }, response, next) => {
 	response.locals.caseId = params.caseId || '';
 	next();
 };
-
-/**
- * Register and validate the caseId and the repId
- *
- * @type {import('express').RequestHandler<*, *, *, *, *>}
- */
-export const registerRepsParams = ({ params }, response, next) => {
-	response.locals.caseId = params.caseId || '';
-	response.locals.repId = params.repId || '';
-
-	next();
-};
